Use GetUser param decorator to ensure user can only delete and update their own account

+++ src/users/users.resolver.ts
@@ -2,8 +2,9 @@ import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
 import { UsersService } from './users.service';
 import { User } from './entities/user.entity';
 import { UpdateUserInput } from './dto/update-user.input';
-import { UseGuards } from '@nestjs/common';
+import { UnauthorizedException, UseGuards } from '@nestjs/common';
 import { AuthenticatedGuard } from 'src/guards/authenticated.guard';
+import { GetUser } from 'src/decorators/user.decorator';
 
 @Resolver(() => User)
 @UseGuards(AuthenticatedGuard)
@@ -20,16 +21,22 @@ export class UsersResolver {
     return this.usersService.findOneById(id);
   }
 
-  @Mutation(() => User)
-  removeUser(@Args('id', { type: () => Int }) id: number) {
+  @Mutation(() => Boolean)
+  removeUser(
+    @Args('id', { type: () => Int }) id: number,
+    @GetUser() user: User,
+  ) {
+    if (id !== user.id) throw new UnauthorizedException();
     return this.usersService.remove(id);
   }
 
-  @Mutation(() => User)
+  @Mutation(() => Boolean)
   updateUser(
     @Args('updateUserInput') updateUserInput: UpdateUserInput,
     @Args('id') id: number,
+    @GetUser() user: User,
   ) {
+    if (id !== user.id) throw new UnauthorizedException();
     return this.usersService.update(id, updateUserInput);
   }
 }
